requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
asgiref==3.7.2
async-timeout==4.0.2
attrs==22.2.0
boto3==1.26.84
botocore==1.29.84
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Django==4.2.2
django-cors-headers==4.1.0
django-extensions==3.2.3
djangorestframework==3.14.0
drf-yasg==1.21.6
fastapi==0.92.0
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
inflection==0.5.1
jmespath==1.0.1
multidict==6.0.4
openai==0.27.0
packaging==23.1
Pillow==9.5.0
pydantic==1.10.5
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.16
sqlparse==0.4.4
starlette==0.25.0
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
yarl==1.8.2
